refactor(DF-54): 모달 컴포넌트 리팩토링
