+++ src/components/modal/Modal.tsx
@@ -1,67 +1,90 @@
-import styled from "styled-components";
+import React from "react";
+import S from "./modal.styles";
+import { theme } from "src/styles/colors/theme";
 import { ReactNode } from "react";
+import {
+  MdWarning,
+  MdCheckCircle,
+  MdInfoOutline,
+  MdDirectionsWalk,
+} from "react-icons/md";
 
-interface ModalProps {
+interface ConfirmationModalProps {
   isOpen: boolean;
   onClose: () => void;
-  children: ReactNode;
-  width?: string;
-  height?: string;
+  onConfirm: () => void;
+  message: ReactNode;
+  cancelText?: string;
+  confirmText?: string;
+  modalType?:
+    | "stop"
+    | "done"
+    | "back"
+    | "delete"
+    | "default"
+    | "location"
+    | "secession";
+  mode?: "create" | "follow";
 }
 
-const ModalOverlay = styled.div<{ isOpen: boolean }>`
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  // background-color: rgba(0, 0, 0, 0.5);
-  display: ${({ isOpen }) => (isOpen ? "flex" : "none")};
-  justify-content: center;
-  align-items: center;
-  z-index: 9999;
-`;
-
-const ModalContainer = styled.div<{ width?: string; height?: string }>`
-  background-color: white;
-  border-radius: 12px;
-  padding: 30px;
-  width: ${({ width }) => width || "85%"};
-  max-width: ${({ width }) => width || "400px"};
-  height: ${({ height }) => height || "auto"};
-  max-height: 90vh;
-  overflow: auto;
-  position: relative;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-
-  /* 스크롤바 숨기기 */
-  &::-webkit-scrollbar {
-    display: none;
-  }
-  -ms-overflow-style: none;
-  scrollbar-width: none;
-
-  /* 태블릿 환경 */
-  @media screen and (min-width: 700px) {
-    padding: 35px;
-    border-radius: 16px;
-    max-width: ${({ width }) => width || "500px"};
-  }
-
-  /* 큰 태블릿 및 노트북 */
-  @media screen and (min-width: 1024px) {
-    max-width: ${({ width }) => width || "600px"};
-    padding: 40px;
-  }
-`;
-
-export default function Modal({
+const Modal = ({
   isOpen,
   onClose,
-  children,
-  width,
-  height,
-}: ModalProps) {
+  onConfirm,
+  message,
+  cancelText = "취소",
+  confirmText = "확인",
+  modalType = "default",
+  mode = "create",
+}: ConfirmationModalProps) => {
+  const getModalContent = () => {
+    switch (modalType) {
+      case "location":
+        return {
+          icon: <MdWarning size={52} color={theme.Red300} />,
+          title: "위치 접근 권한 필요",
+        };
+      case "stop":
+        return {
+          icon: <MdWarning size={52} color={theme.Red300} />,
+          title: "산책 조건 미달",
+        };
+      case "done":
+        if (mode === "create") {
+          return {
+            icon: <MdCheckCircle size={52} color={theme.Green500} />,
+            title: "산책로 등록",
+          };
+        } else {
+          return {
+            icon: <MdDirectionsWalk size={52} color={theme.Following} />,
+            title: "이용 완료",
+          };
+        }
+      case "back":
+        return {
+          icon: <MdInfoOutline size={52} color={theme.Following} />,
+        };
+      case "delete":
+        return {
+          icon: <MdWarning size={52} color={theme.Red300} />,
+          title: "산책로 삭제",
+        };
+      case "secession":
+        return {
+          icon: <MdWarning size={52} color={theme.Red300} />,
+          title: "서비스 탈퇴",
+        };
+      default:
+        return {
+          icon: <MdInfoOutline size={52} color={theme.Following} />,
+          title: "알림",
+        };
+    }
+  };
+
+  const { icon, title } = getModalContent();
+
   const handleOverlayClick = (e: React.MouseEvent<HTMLDivElement>) => {
     if (e.target === e.currentTarget) {
       onClose();
@@ -69,10 +92,26 @@ export default function Modal({
   };
 
   return (
-    <ModalOverlay isOpen={isOpen} onClick={handleOverlayClick}>
-      <ModalContainer width={width} height={height}>
-        {children}
-      </ModalContainer>
-    </ModalOverlay>
+    <S.ModalOverlay isOpen={isOpen} onClick={handleOverlayClick}>
+      <S.ModalContainer>
+        <S.ModalContent>
+          <S.IconContainer>{icon}</S.IconContainer>
+
+          <S.TextContainer>
+            {title && <S.TitleText>{title}</S.TitleText>}
+            <S.ModalText>{message}</S.ModalText>
+          </S.TextContainer>
+
+          <S.ButtonContainer>
+            <S.Button onClick={onClose}>{cancelText}</S.Button>
+            <S.Button isConfirm onClick={onConfirm}>
+              {confirmText}
+            </S.Button>
+          </S.ButtonContainer>
+        </S.ModalContent>
+      </S.ModalContainer>
+    </S.ModalOverlay>
   );
-}
+};
+
+export default Modal;
